+++ payments/src/sdk/models/shared/bulkpayableitemcreate.ts
@@ -40,6 +40,10 @@ export class BulkPayableItemCreate extends SpeakeasyBase {
     bulkPayableItemReference?: string;
 
     @SpeakeasyMetadata()
+    @Expose({ name: "clientId" })
+    clientId?: string;
+
+    @SpeakeasyMetadata()
     @Expose({ name: "collaboratorEmail" })
     collaboratorEmail?: string;