+++ payments/src/sdk/sdk.ts
@@ -69,9 +69,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "1.0.0";
-    sdkVersion = "2.1.1";
+    sdkVersion = "2.1.2";
     genVersion = "2.223.3";
-    userAgent = "speakeasy-sdk/typescript 2.1.1 2.223.3 1.0.0 @wingspan/payments";
+    userAgent = "speakeasy-sdk/typescript 2.1.2 2.223.3 1.0.0 @wingspan/payments";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);